Check error code when test is supposed to fail
Tests that are supposed to fail assume that no RDF file is generated.
While that makes sense, there is also an argument for files being
written in batches. Deletion of a file may be expensive. One should
also look whether the execution of a command returned a non-zero value
indicating an error.